Rename seatbelt_common crate to macos_sandbox_common
Rename the crate, module, and file from seatbelt-based naming to
macos_sandbox-based naming for consistency with the containment backend
name. Keep SeatbeltScriptRunner struct name since it specifically refers
to the sandbox-exec (Seatbelt) execution mode.

- Crate: seatbelt_common -> macos_sandbox_common
- Module: seatbelt_runner -> macos_sandbox_runner
- Function: run_seatbelt() -> run_macos_sandbox()
- Docs: clarify "Apple's Seatbelt" when referring to the technology
- Update SDK comment and CLI about text
