+++ src/Cargo.toml
@@ -12,7 +12,7 @@ members = [
     "lxc_common",
     "wslc_common",
     "isolation_session_bindings",
-    "seatbelt_common",
+    "macos_sandbox_common",
     "mxc_darwin",
     "generated/base_container_specification",
     "mxc_diagnostic_console",
